refactor views and enable auth
