+++ content/templates/base.html
@@ -26,6 +26,26 @@
           Add Content
         </div>
       </a>
+      {% if not user.is_authenticated %}
+      <a href="{% url 'profile' %}">
+        <div>
+          Login
+        </div>
+      </a>
+      {% else %}
+      <a href="{% url 'profile' %}">
+        <div>
+          Profile
+        </div>
+      </a>
+      {% endif %}
+      {% if user.is_authenticated %}
+      <a href="{% url 'logout' %}">
+        <div>
+          Logout
+        </div>
+      </a>
+      {% endif %}
     </nav>
     <div class="container">
     {% block content %}
