+++ content/templates/profile.html
@@ -1,0 +1,40 @@
+{% extends 'base.html' %}
+{% block content %}
+<h4>{{ user }}</h4>
+<p>{{ user.first_name }} {{ user.last_name }}</p>
+<h4>Your Artists:</h4>
+<table class="table-list">
+  <tbody>
+  {% for obj in user.artist_set.all %}
+
+      <tr class="table-row clickable" onclick="document.location = '{% url 'artist-detail' obj.id %}'">
+        <td><a class="basic-link" href="{% url 'artist-detail' obj.id %}">{{ obj.name }}</a></td>
+        <td>{% if obj.has_other_aliases %}({{ obj.other_aliases|join:", " }}){% endif %}</td>
+    </tr>
+  {% endfor %}
+  </tbody>
+</table>
+<h4>Your Albums:</h4>
+<table class="table-list">
+  <tbody>
+  {% for album in user.album_set.all %}
+    <tr class="table-row clickable" onclick="document.location = '{% url 'album-detail' album.id %}'">
+      <td><a class="basic-link" href="{% url 'album-detail' album.id %}">{{ album.title }}</a></td>
+      <td>{{ album.contributors.first.artist.name }}</td>
+    </tr>
+  {% endfor %}
+  </tbody>
+</table>
+<h4>Your Tracks:</h4>
+<table class="table-list">
+  <tbody>
+  {% for obj in user.track_set.all %}
+
+      <tr class="table-row clickable" onclick="document.location = '{% url 'track-detail' obj.id %}'">
+        <td><a class="basic-link" href="{% url 'track-detail' obj.id %}">{{ obj.title }}</a></td>
+        <td>{{ obj.contributor.artist.name }}</td>
+    </tr>
+  {% endfor %}
+  </tbody>
+</table>
+{% endblock content %}
