+++ content/templates/registration/login.html
@@ -1,0 +1,8 @@
+{% extends 'base.html' %}
+{% block content %}
+<form method="POST">
+{% csrf_token %}
+{{ form }}
+<input type="submit" value="Login"/>
+</form>
+{% endblock content %}
