+++ content/views/login.py
@@ -1,0 +1,13 @@
+from django.contrib.auth.views import LoginView, LogoutView
+
+
+class WayloLoginView(LoginView):
+    """The login view
+    """
+
+
+class WayloLogoutView(LogoutView):
+    """The logout view
+    """
+
+    next_page = "/"
