+++ content/views/profile.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render
+from django.contrib.auth.decorators import login_required
+
+
+@login_required
+def user_profile_view(request):
+    return render(request, "profile.html")
